chore: minor code quality improved

+++ common/src/main/java/com/stevekung/fishofthieves/entity/animal/Battlegill.java
@@ -125,7 +125,7 @@ public class Battlegill extends AbstractSchoolingThievesFish<BattlegillVariant>
     @Override
     public double getMeleeAttackRangeSqr(LivingEntity entity)
     {
-        return 1.0 + (double) entity.getBbWidth() * 2.0;
+        return 1.0 + entity.getBbWidth() * 2.0;
     }
 
     @Override
